+++ src/Phan/Language/Type.php
@@ -54,9 +54,6 @@ use Phan\Language\Type\StringType;
 use Phan\Language\Type\TemplateType;
 use Phan\Language\Type\TrueType;
 use Phan\Language\Type\VoidType;
-use Phan\Library\None;
-use Phan\Library\Option;
-use Phan\Library\Some;
 use Phan\Library\Tuple5;
 use function count;
 use function explode;
@@ -529,96 +526,6 @@ class Type
     }
 
     /**
-     * If the $name is a reserved constant, then returns the NativeType for that name
-     * Otherwise, this returns null.
-     * @return Option<NativeType>
-     */
-    public static function fromReservedConstantName(string $name) : Option
-    {
-        static $lookup;
-        if ($lookup === null) {
-            $lookup = self::createReservedConstantNameLookup();
-        }
-        $result = $lookup[\strtoupper(\ltrim($name, '\\'))] ?? null;
-        if (isset($result)) {
-            return new Some($result);
-        }
-        return new None();
-    }
-
-    /**
-     * @return array<string,NativeType> a map from the **uppercase** reserved constant name to the subclass of NativeType for that constant.
-     * Uses the constants and types from https://secure.php.net/manual/en/reserved.constants.php
-     */
-    private static function createReservedConstantNameLookup() : array
-    {
-        $false  = FalseType::instance(false);
-        // $float  = FloatType::instance(false);
-        $int    = IntType::instance(false);
-        $null   = NullType::instance(false);
-        $string = StringType::instance(false);
-        $true   = TrueType::instance(false);
-
-        return [
-            'PHP_VERSION'           => $string,
-            'PHP_MAJOR_VERSION'     => $int,
-            'PHP_MINOR_VERSION'     => $int,
-            'PHP_RELEASE_VERSION'   => $int,
-            'PHP_VERSION_ID'        => $int,
-            'PHP_EXTRA_VERSION'     => $string,
-            'PHP_ZTS'               => $int,
-            'PHP_MAXPATHLEN'        => $int,
-            'PHP_OS'                => $string,
-            'PHP_OS_FAMILY'         => $string,
-            'PHP_SAPI'              => $string,
-            'PHP_EOL'               => $string,
-            'PHP_INT_MAX'           => Type::fromObject(\PHP_INT_MAX),
-            'PHP_INT_MIN'           => Type::fromObject(\PHP_INT_MIN),  // since 7.0.0
-            'PHP_INT_SIZE'          => Type::fromObject(\PHP_INT_SIZE),  // since 7.0.0
-            //'PHP_FLOAT_DIG'         => $int,  // since 7.2.0
-            //'PHP_FLOAT_EPSILON'     => $float,  // since 7.2.0
-            //'PHP_FLOAT_MIN'         => $int, // since 7.2.0
-            //'PHP_FLOAT_MAX'         => $int, // since 7.2.0
-            'DEFAULT_INCLUDE_PATH'  => $string,
-            'PEAR_INSTALL_DIR'      => $string,
-            'PHP_EXTENSION_DIR'     => $string,
-            'PEAR_EXTENSION_DIR'    => $string,
-            'PHP_PREFIX'            => $string,
-            'PHP_BINDIR'            => $string,
-            'PHP_BINARY'            => $string,
-            'PHP_MANDIR'            => $string,
-            'PHP_LIBDIR'            => $string,
-            'PHP_DATADIR'           => $string,
-            'PHP_SYSCONFDIR'        => $string,
-            'PHP_LOCALSTATEDIR'     => $string,
-            'PHP_CONFIG_FILE_PATH'  => $string,
-            'PHP_CONFIG_FILE_SCAN_DIR' => $string,
-            'PHP_SHLIB_SUFFIX'      => $string,
-            //'PHP_FD_SETSIZE'            => $int,  // 7.2.0 TODO: web page documentation is wrong, says string.
-            'E_ERROR'               => $int,
-            'E_WARNING'             => $int,
-            'E_PARSE'               => $int,
-            'E_NOTICE'              => $int,
-            'E_CORE_ERROR'          => $int,
-            'E_CORE_WARNING'        => $int,
-            'E_COMPILE_ERROR'       => $int,
-            'E_COMPILE_WARNING'     => $int,
-            'E_USER_ERROR'          => $int,
-            'E_USER_WARNING'        => $int,
-            'E_USER_NOTICE'         => $int,
-            'E_DEPRECATED'          => $int,
-            'E_USER_DEPRECATED'     => $int,
-            'E_ALL'                 => $int,
-            'E_STRICT'              => $int,
-            '__COMPILER_HALT_OFFSET__' => $int,
-            '__LINE__'              => $int,
-            'TRUE'                  => $true,
-            'FALSE'                 => $false,
-            'NULL'                  => $null,
-        ];
-    }
-
-    /**
      * @param mixed $object
      * @return Type
      * Get a type for the given object. Equivalent to Type::fromObject($object)->asNonLiteralType()
@@ -643,12 +550,11 @@ class Type
     }
 
     /**
-     * @param mixed $object
-     * @return NativeType
      * Get a type for the given object
+     * @param mixed $object
      * @throws AssertionError if the type was unexpected
      */
-    public static function fromObject($object) : NativeType
+    public static function fromObject($object) : Type
     {
         switch (\gettype($object)) {
             case 'integer':
@@ -662,7 +568,9 @@ class Type
             case 'double':
                 return FloatType::instance(false);
             case 'object':
-                return ObjectType::instance(false);
+                // TODO: Figure out why this is a false positive
+                // @phan-suppress-next-line PhanTypeMismatchArgumentInternal, PhanThrowTypeMismatchForCall
+                return Type::fromFullyQualifiedString('\\' . \get_class($object));
             case 'boolean':
                 return $object ? TrueType::instance(false) : FalseType::instance(false);
             case 'array':
@@ -672,6 +580,42 @@ class Type
             default:
                 throw new \AssertionError("Unknown type " . \gettype($object));
         }
+    }
+
+    /**
+     * Get a type for the given object.
+     * If $object is an array, return an ArrayShapeType (with nested fields) instead of an ArrayType
+     * @param mixed $object
+     * @throws AssertionError if the type was unexpected
+     */
+    public static function fromObjectExtended($object) : Type
+    {
+        if (\is_array($object)) {
+            return self::fromArray($object);
+        }
+        return self::fromObject($object);
+    }
+
+    /**
+     * Get a type for the given array as an array shape, recursively.
+     * @param array<mixed,mixed> $array
+     * @throws AssertionError if the type was unexpected
+     */
+    public static function fromArray(array $array) : ArrayShapeType
+    {
+        return ArrayShapeType::fromFieldTypes(
+            array_map(
+                /**
+                 * @param mixed $value
+                 * @return UnionType
+                 */
+                static function ($value) {
+                    return self::fromObjectExtended($value)->asUnionType();
+                },
+                $array
+            ),
+            false
+        );
     }
 
     /**